[front] chore: Migrate front MCP server
